format: add utility functions to lib file
